% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{calculate_distance}
\alias{calculate_distance}
\title{Calculate Distance}
\usage{
calculate_distance(lat1, lon1, lat2, lon2)
}
\arguments{
\item{lat1}{Latitude of the first point.}

\item{lon1}{Longitude of the first point.}

\item{lat2}{Latitude of the second point.}

\item{lon2}{Longitude of the second point.}
}
\value{
Distance in miles.
}
\description{
This function calculates the distance between two points specified by their
latitude and longitude.
}
