% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{get_census_code_vec}
\alias{get_census_code_vec}
\title{Get Census Code Vector}
\usage{
get_census_code_vec(lat, lon)
}
\arguments{
\item{lat}{A vector of latitudes.}

\item{lon}{A vector of longitudes.}
}
\value{
A vector of census codes corresponding to the input latitude and longitude pairs.
If a latitude or longitude within a pair is NA the census code will be NA
}
\description{
This function calculates the census code for a vector of latitude and longitude pairs.
}
